+++ database/migrations/2020_01_29_162459_create_invites_table.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
@@ -6,7 +6,7 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    public function up() :void
+    public function up(): void
     {
         $table_name = config('invite-codes.tables.invites_table', 'invites');