+++ src/Exceptions/UserLoggedOutException.php
@@ -1,7 +1,5 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Junges\InviteCodes\Exceptions;
 
-class UserLoggedOutException extends InviteCodesException
-{
-}
+class UserLoggedOutException extends InviteCodesException {}