+++ src/InviteCodesServiceProvider.php
@@ -1,10 +1,11 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Junges\InviteCodes;
 
 use Illuminate\Support\ServiceProvider;
 use Junges\InviteCodes\Console\Commands\DeleteExpiredInvitesCommand;
 use Junges\InviteCodes\Contracts\InviteCodesFactory;
+use Override;
 
 final class InviteCodesServiceProvider extends ServiceProvider
 {
@@ -13,6 +14,13 @@ final class InviteCodesServiceProvider extends ServiceProvider
         $this->publishesConfig();
         $this->loadMigrations();
         $this->loadCommands();
+    }
+
+    /** Register any application services. */
+    #[Override]
+    public function register(): void
+    {
+        $this->app->bind(InviteCodesFactory::class, Factory::class);
     }
 
     /** Load and publishes the package configuration file. */
@@ -45,11 +53,5 @@ final class InviteCodesServiceProvider extends ServiceProvider
                 DeleteExpiredInvitesCommand::class,
             ]);
         }
-    }
-
-    /** Register any application services. */
-    public function register(): void
-    {
-        $this->app->bind(InviteCodesFactory::class, Factory::class);
     }
 }